pull error checker out into a separate file
