+++ test/test_error_checker.rb
@@ -1,13 +1,12 @@
 require_relative "helper"
 
 class TestErrorChecker < MiniTest::Test
-
   def test_check_with_status_code_412
     response = mock
     response.stubs(:status_code).returns(412)
     response.stubs(:body).returns('junk')
     assert_raises Rets::HttpError do
-      Rets::Client::ErrorChecker.check(response)
+      Rets::Parser::ErrorChecker.check(response)
     end
   end
 
@@ -18,7 +17,7 @@ class TestErrorChecker < MiniTest::Test
     response.stubs(:ok?).returns(false)
     response.stubs(:body).returns('')
     assert_raises Rets::AuthorizationFailure do
-      Rets::Client::ErrorChecker.check(response)
+      Rets::Parser::ErrorChecker.check(response)
     end
   end
 
@@ -29,7 +28,7 @@ class TestErrorChecker < MiniTest::Test
     response.stubs(:ok?).returns(false)
     response.stubs(:body).returns(HTML_AUTH_FAILURE)
     assert_raises Rets::AuthorizationFailure do
-      Rets::Client::ErrorChecker.check(response)
+      Rets::Parser::ErrorChecker.check(response)
     end
   end
 
@@ -40,8 +39,7 @@ class TestErrorChecker < MiniTest::Test
     response.stubs(:ok?).returns(false)
     response.stubs(:body).returns(XHTML_AUTH_FAILURE)
     assert_raises Rets::AuthorizationFailure do
-      Rets::Client::ErrorChecker.check(response)
+      Rets::Parser::ErrorChecker.check(response)
     end
   end
-
 end
